v 0.6.0 Full cruds (Users, Statuses, Task, Labels) with task_used_check before deleting and full tests

+++ task_manager/mixins.py
@@ -2,7 +2,6 @@ from django.shortcuts import redirect
 from django.utils.translation import gettext as _
 from django.urls import reverse
 from django.contrib import messages
-from django.db import IntegrityError
 
 
 class LoginRequiredRedirectMixin:
@@ -46,29 +45,28 @@ class TaskUnusedRequaredDeletionMixin:
     If is used - prevent delete (redirect with message).
     If isn't used - delete (redirect with message)
     Attrs: message_used_object: str, url_name_object_used,
-    message_success: str, url_name_success: str,
-    You can set only url_name:str for both redirects
+    message_success: str, url_name_success: str
     '''
-    url_name = 'main_page'
-    message_used_object = _('Unable to delete! This object is bounded!')
-    url_name_object_used = url_name
+    message_used_object = _('Unable to delete because it is used in task!')
+    url_name_object_used = "main_page"
     message_success = _("Object is deleted successfully!")
-    url_name_success = url_name
+    url_name_success = "main_page"
+
+    def is_task_used(self):
+        model_name = self.object.__class__.__name__
+        if model_name in ["Label", "Status"]:
+            return True if self.object.task_set.all() else False
+        # User had two related managers, both differ from task_set
+        if self.object.task_creator_set.all():
+            return True
+        if self.object.task_executor_set.all():
+            return True
+        return False
 
     def form_valid(self, form):
-        try:
-            self.object.delete()
-            messages.success(self.request, self.message_success)
-        except IntegrityError:
+        if self.is_task_used():
             messages.warning(self.request, self.message_used_object)
-            redirect(self.url_name_object_used)
+            return redirect(self.url_name_object_used)
+        messages.success(self.request, self.message_success)
+        self.object.delete()
         return redirect(self.url_name_success)
-
-    # def form_valid(self, form):
-    #     print('task_set: ', self.object.task_set)
-    #     if self.object.task_set.all():
-    #         messages.warning(self.request, self.message_used_object)
-    #         redirect(self.url_name_object_used)
-    #     self.object.delete()
-    #     messages.success(self.request, self.message_success)
-    #     return redirect(self.url_name_success)
